CSCE704-Proj/defender/docker-requirements.txt
# pip install git+https://github.com/endgameinc/ember.git
flask>=1.1.2
gevent>=1.4.0
envparse
annoy
numpy
tqdm
pandas
scipy
scikit-learn==0.24.1
pefile
Pillow
numpy
matplotlib
tensorflow==2.18.0